% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticate_googlesheets}
\alias{authenticate_googlesheets}
\title{Authenticate in order to use Google Sheets}
\usage{
authenticate_googlesheets()
}
\value{

}
\description{
Authenticate in order to use Google Sheets
}
