% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticate_twitter}
\alias{authenticate_twitter}
\title{Generate Twitter token}
\usage{
authenticate_twitter()
}
\value{
A data frame twitter_token that can be used to authenticate with Twitter
}
\description{
Generate Twitter token
}
