% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweet_data.R
\name{get_tweet_data}
\alias{get_tweet_data}
\title{Get tweet data and post to Google Sheet}
\usage{
get_tweet_data()
}
\value{
Posts data on Google Sheet
}
\description{
Get tweet data and post to Google Sheet
}
