% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_gif.R
\name{pdf_to_gif}
\alias{pdf_to_gif}
\title{Convert PDF to Animated GIF}
\usage{
pdf_to_gif(pdf_location, gif_location, density = 50)
}
\arguments{
\item{pdf_location}{}
}
\value{

}
\description{
Convert PDF to Animated GIF
}
