% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rru_tweet.R
\name{rru_tweet}
\alias{rru_tweet}
\title{Get tweets from Google Sheets, post one, then put data back on Google Sheet}
\usage{
rru_tweet(sheet_name)
}
\arguments{
\item{sheet_name}{}
}
\value{

}
\description{
Get tweets from Google Sheets, post one, then put data back on Google Sheet
}
